% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brewer.R
\name{brewer}
\alias{brewer}
\alias{brewer_accent}
\alias{brewer_blues}
\alias{brewer_brbg}
\alias{brewer_bugn}
\alias{brewer_bupu}
\alias{brewer_dark2}
\alias{brewer_gnbu}
\alias{brewer_greens}
\alias{brewer_greys}
\alias{brewer_oranges}
\alias{brewer_orrd}
\alias{brewer_paired}
\alias{brewer_pastel1}
\alias{brewer_pastel2}
\alias{brewer_piyg}
\alias{brewer_prgn}
\alias{brewer_pubu}
\alias{brewer_pubugn}
\alias{brewer_puor}
\alias{brewer_purd}
\alias{brewer_purples}
\alias{brewer_rdbu}
\alias{brewer_rdgy}
\alias{brewer_rdpu}
\alias{brewer_rdylbl}
\alias{brewer_rdylgn}
\alias{brewer_reds}
\alias{brewer_set1}
\alias{brewer_set2}
\alias{brewer_set3}
\alias{brewer_spectral}
\alias{brewer_ylgn}
\alias{brewer_ylgnbu}
\alias{brewer_ylorbr}
\alias{brewer_ylorrd}
\title{ColorBrewer palettes}
\usage{
brewer_blues(n)

brewer_bugn(n)

brewer_bupu(n)

brewer_gnbu(n)

brewer_greens(n)

brewer_greys(n)

brewer_oranges(n)

brewer_orrd(n)

brewer_pubu(n)

brewer_pubugn(n)

brewer_purd(n)

brewer_purples(n)

brewer_rdpu(n)

brewer_reds(n)

brewer_ylgn(n)

brewer_ylgnbu(n)

brewer_ylorbr(n)

brewer_ylorrd(n)

brewer_brbg(n)

brewer_piyg(n)

brewer_prgn(n)

brewer_puor(n)

brewer_rdbu(n)

brewer_rdgy(n)

brewer_rdylbl(n)

brewer_rdylgn(n)

brewer_spectral(n)

brewer_accent(n)

brewer_dark2(n)

brewer_paired(n)

brewer_pastel1(n)

brewer_pastel2(n)

brewer_set1(n)

brewer_set2(n)

brewer_set3(n)
}
\description{
These functions provide a unified interface to the ColorBrewer palettes.
}
\details{
The palette names begin with 'brewer' to make it easier to use auto-completion.
For consistency with the rest of this package, the palette names use lower-case letters.
}
\examples{

# Qualitative

pal.test(brewer_blues)
pal.test(brewer_bugn)
pal.test(brewer_bupu)
pal.test(brewer_gnbu)
pal.test(brewer_greens)
pal.test(brewer_greys)
pal.test(brewer_oranges)
pal.test(brewer_orrd)
pal.test(brewer_pubu)
pal.test(brewer_pubugn)
pal.test(brewer_purd)
pal.test(brewer_purples)
pal.test(brewer_rdpu)
pal.test(brewer_reds)
pal.test(brewer_ylgn)
pal.test(brewer_ylgnbu)
pal.test(brewer_ylorbr)
pal.test(brewer_ylorrd)

# Diverging, max 11

pal.test(brewer_brbg)
pal.test(brewer_piyg)
pal.test(brewer_prgn)
pal.test(brewer_puor)
pal.test(brewer_rdbu)
pal.test(brewer_rdgy)
pal.test(brewer_rdylbl)
pal.test(brewer_rdylgn)
pal.test(brewer_spectral)

# Qualtitative
# These are weird...don't do this

pal.test(brewer_accent)
pal.test(brewer_dark2)
pal.test(brewer_paired)
pal.test(brewer_pastel1)
pal.test(brewer_pastel2)
pal.test(brewer_set1)
pal.test(brewer_set2)
pal.test(brewer_set3)

pal.bands(c('brewer_accent','brewer_dark2','brewer_paired','brewer_pastel1','brewer_pastel2','brewer_set1','brewer_set2','brewer_set3'))

}

