% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaps.R
\name{kovesi}
\alias{kovesi}
\alias{kovesi.cyclic_grey_15_85_c0}
\alias{kovesi.cyclic_grey_15_85_c0_s25}
\alias{kovesi.cyclic_mrybm_35_75_c68}
\alias{kovesi.cyclic_mrybm_35_75_c68_s25}
\alias{kovesi.cyclic_mygbm_30_95_c78}
\alias{kovesi.cyclic_mygbm_30_95_c78_s25}
\alias{kovesi.cyclic_wrwbw_40_90_c42}
\alias{kovesi.cyclic_wrwbw_40_90_c42_s25}
\alias{kovesi.diverging_isoluminant_cjm_75_c23}
\alias{kovesi.diverging_isoluminant_cjm_75_c24}
\alias{kovesi.diverging_isoluminant_cjo_70_c25}
\alias{kovesi.diverging_linear_bjr_30_55_c53}
\alias{kovesi.diverging_linear_bjy_30_90_c45}
\alias{kovesi.diverging_rainbow_bgymr_45_85_c67}
\alias{kovesi.diverging_bkr_55_10_c35}
\alias{kovesi.diverging_bky_60_10_c30}
\alias{kovesi.diverging_bwr_40_95_c42}
\alias{kovesi.diverging_bwr_55_98_c37}
\alias{kovesi.diverging_cwm_80_100_c22}
\alias{kovesi.diverging_gkr_60_10_c40}
\alias{kovesi.diverging_gwr_55_95_c38}
\alias{kovesi.diverging_gwv_55_95_c39}
\alias{kovesi.isoluminant_cgo_70_c39}
\alias{kovesi.isoluminant_cgo_80_c38}
\alias{kovesi.isoluminant_cm_70_c39}
\alias{kovesi.linear_bgy_10_95_c74}
\alias{kovesi.linear_bgyw_15_100_c67}
\alias{kovesi.linear_bgyw_15_100_c68}
\alias{kovesi.linear_blue_5_95_c73}
\alias{kovesi.linear_blue_95_50_c20}
\alias{kovesi.linear_bmw_5_95_c86}
\alias{kovesi.linear_bmw_5_95_c89}
\alias{kovesi.linear_bmy_10_95_c71}
\alias{kovesi.linear_bmy_10_95_c78}
\alias{kovesi.linear_gow_60_85_c27}
\alias{kovesi.linear_gow_65_90_c35}
\alias{kovesi.linear_green_5_95_c69}
\alias{kovesi.linear_grey_0_100_c0}
\alias{kovesi.linear_grey_10_95_c0}
\alias{kovesi.linear_kry_5_95_c72}
\alias{kovesi.linear_kry_5_98_c75}
\alias{kovesi.linear_kryw_5_100_c64}
\alias{kovesi.linear_kryw_5_100_c67}
\alias{kovesi.linear_ternary_blue_0_44_c57}
\alias{kovesi.linear_ternary_green_0_46_c42}
\alias{kovesi.linear_ternary_red_0_50_c52}
\alias{kovesi.rainbow_bgyr_35_85_c72}
\alias{kovesi.rainbow}
\alias{kovesi.rainbow_bgyr_35_85_c73}
\alias{kovesi.rainbow_bgyrm_35_85_c69}
\alias{kovesi.rainbow_bgyrm_35_85_c71}
\title{Peter Kovesi's perceptually uniform colormaps}
\usage{
kovesi.cyclic_grey_15_85_c0(n)

kovesi.cyclic_grey_15_85_c0_s25(n)

kovesi.cyclic_mrybm_35_75_c68(n)

kovesi.cyclic_mrybm_35_75_c68_s25(n)

kovesi.cyclic_mygbm_30_95_c78(n)

kovesi.cyclic_mygbm_30_95_c78_s25(n)

kovesi.cyclic_wrwbw_40_90_c42(n)

kovesi.cyclic_wrwbw_40_90_c42_s25(n)

kovesi.diverging_isoluminant_cjm_75_c23(n)

kovesi.diverging_isoluminant_cjm_75_c24(n)

kovesi.diverging_isoluminant_cjo_70_c25(n)

kovesi.diverging_linear_bjr_30_55_c53(n)

kovesi.diverging_linear_bjy_30_90_c45(n)

kovesi.diverging_rainbow_bgymr_45_85_c67(n)

kovesi.diverging_bkr_55_10_c35(n)

kovesi.diverging_bky_60_10_c30(n)

kovesi.diverging_bwr_40_95_c42(n)

kovesi.diverging_bwr_55_98_c37(n)

kovesi.diverging_cwm_80_100_c22(n)

kovesi.diverging_gkr_60_10_c40(n)

kovesi.diverging_gwr_55_95_c38(n)

kovesi.diverging_gwv_55_95_c39(n)

kovesi.isoluminant_cgo_70_c39(n)

kovesi.isoluminant_cgo_80_c38(n)

kovesi.isoluminant_cm_70_c39(n)

kovesi.linear_bgy_10_95_c74(n)

kovesi.linear_bgyw_15_100_c67(n)

kovesi.linear_bgyw_15_100_c68(n)

kovesi.linear_blue_5_95_c73(n)

kovesi.linear_blue_95_50_c20(n)

kovesi.linear_bmw_5_95_c86(n)

kovesi.linear_bmw_5_95_c89(n)

kovesi.linear_bmy_10_95_c71(n)

kovesi.linear_bmy_10_95_c78(n)

kovesi.linear_gow_60_85_c27(n)

kovesi.linear_gow_65_90_c35(n)

kovesi.linear_green_5_95_c69(n)

kovesi.linear_grey_0_100_c0(n)

kovesi.linear_grey_10_95_c0(n)

kovesi.linear_kry_5_95_c72(n)

kovesi.linear_kry_5_98_c75(n)

kovesi.linear_kryw_5_100_c64(n)

kovesi.linear_kryw_5_100_c67(n)

kovesi.linear_ternary_blue_0_44_c57(n)

kovesi.linear_ternary_green_0_46_c42(n)

kovesi.linear_ternary_red_0_50_c52(n)

kovesi.rainbow_bgyr_35_85_c72(n)

kovesi.rainbow(n)

kovesi.rainbow_bgyr_35_85_c73(n)

kovesi.rainbow_bgyrm_35_85_c69(n)

kovesi.rainbow_bgyrm_35_85_c71(n)
}
\arguments{
\item{n}{The number of colors to display for palette functions.}
}
\value{
A vector of colors.
}
\description{
Peter Kovesi's perceptually uniform colormaps
}
\details{
All colormaps are named using Peter Kovesi's naming scheme:
<category>_<huesequence>_<lightnessrange>_c<meanchroma>_s<colorshift>

Note: \code{kovesi.rainbow} is another name for \code{rainbow_bgyr_35_85_c72}.
}
\examples{

if(FALSE){
pal.bands(kovesi.cyclic_grey_15_85_c0, kovesi.cyclic_grey_15_85_c0_s25,
kovesi.cyclic_mrybm_35_75_c68, kovesi.cyclic_mrybm_35_75_c68_s25,
kovesi.cyclic_mygbm_30_95_c78, kovesi.cyclic_mygbm_30_95_c78_s25,
kovesi.cyclic_wrwbw_40_90_c42, kovesi.cyclic_wrwbw_40_90_c42_s25,
kovesi.diverging_isoluminant_cjm_75_c23, kovesi.diverging_isoluminant_cjm_75_c24,
kovesi.diverging_isoluminant_cjo_70_c25, kovesi.diverging_linear_bjr_30_55_c53,
kovesi.diverging_linear_bjy_30_90_c45, kovesi.diverging_rainbow_bgymr_45_85_c67,
kovesi.diverging_bkr_55_10_c35, kovesi.diverging_bky_60_10_c30,
kovesi.diverging_bwr_40_95_c42, kovesi.diverging_bwr_55_98_c37,
kovesi.diverging_cwm_80_100_c22, kovesi.diverging_gkr_60_10_c40,
kovesi.diverging_gwr_55_95_c38, kovesi.diverging_gwv_55_95_c39,
kovesi.isoluminant_cgo_70_c39, kovesi.isoluminant_cgo_80_c38,
kovesi.isoluminant_cm_70_c39, kovesi.linear_bgy_10_95_c74,
kovesi.linear_bgyw_15_100_c67, kovesi.linear_bgyw_15_100_c68,
kovesi.linear_blue_5_95_c73, kovesi.linear_blue_95_50_c20,
kovesi.linear_bmw_5_95_c86, kovesi.linear_bmw_5_95_c89,
kovesi.linear_bmy_10_95_c71, kovesi.linear_bmy_10_95_c78,
kovesi.linear_gow_60_85_c27, kovesi.linear_gow_65_90_c35,
kovesi.linear_green_5_95_c69, kovesi.linear_grey_0_100_c0,
kovesi.linear_grey_10_95_c0, kovesi.linear_kry_5_95_c72,
kovesi.linear_kry_5_98_c75, kovesi.linear_kryw_5_100_c64,
kovesi.linear_kryw_5_100_c67, kovesi.linear_ternary_blue_0_44_c57,
kovesi.linear_ternary_green_0_46_c42, kovesi.linear_ternary_red_0_50_c52,
kovesi.rainbow_bgyr_35_85_c72, kovesi.rainbow_bgyr_35_85_c73,
kovesi.rainbow_bgyrm_35_85_c69, kovesi.rainbow_bgyrm_35_85_c71)
}

}
\references{
Peter Kovesi (2016). CET Perceptually Uniform Colour Maps.
http://peterkovesi.com/projects/colourmaps/

Peter Kovesi (2015). Good Colour Maps: How to Design Them.
Arxiv. https://arxiv.org/abs/1509.03700


https://bokeh.github.io/colorcet/
}
\author{
Colormaps by Peter Kovesi. R code by Kevin Wright.
}
