% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPALS.R
\name{LPALS}
\alias{LPALS}
\title{compute directions using principal asymmetric least squares}
\usage{
LPALS(X, Y, lambda = 1, tau = seq(0.1, 0.9, by = 0.1))
}
\arguments{
\item{X}{A design or model matrix which cannot include categorical variables}

\item{Y}{A continuous response variable}

\item{lambda}{A tuning parameter which must be greater than 0}

\item{tau}{The quantile levels which must be at least two}
}
\value{
Returns the directions and their corresponding eigenvalues, and the coefficients of the expectile estimates
}
\description{
compute directions using principal asymmetric least squares
}
