% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colormaps.R
\name{matplotlib}
\alias{matplotlib}
\alias{magma}
\alias{inferno}
\alias{plasma}
\alias{viridis}
\title{Matplotlib colormaps}
\usage{
magma(n)

inferno(n)

plasma(n)

viridis(n)
}
\arguments{
\item{n}{Number of colors to return}
}
\value{
A vector of colors
}
\description{
Viridis family of colormaps as found in Matplotlib. Designed to be perceptually 
uniform, but generally too dark to be useful.
}
\examples{
pal.bands(magma, inferno, plasma, viridis)

}
\author{
Palettes by Matteo Niccoli. R code by Kevin Wright.
}
