% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matpower.R
\name{matpower}
\alias{matpower}
\title{computes the matrix power}
\usage{
matpower(A, n)
}
\arguments{
\item{A}{A square matrix}

\item{n}{The power of the matrix we want to compute}
}
\value{
Returns a square matrix
}
\description{
computes the matrix power
}
