% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.channels}
\alias{pal.channels}
\title{Show the red, green, blue, gray amount in colors of a palette}
\usage{
pal.channels(pal, n = 150)
}
\arguments{
\item{pal}{A palette function or a vector of colors.}

\item{n}{The number of colors to display for palette functions.}
}
\value{
None
}
\description{
The amount of red, green, blue, and gray in colors are shown.
}
\details{
What to look for:

1. Sequential data should usually be shown with a colormap that is smoothly
increasing in lightness, as shown by the gray line.
}
\examples{
pal.channels(viridis(10))
pal.channels(viridis)

}
\author{
Kevin Wright
}
\references{
None
}

