% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\name{pal.chips}
\alias{pal.chips}
\title{View bands of color palettes}
\usage{
pal.chips(pals, n = 51)
}
\arguments{
\item{pals}{A vector of strings that are the names of
functions that return a color palette.}

\item{n}{The number of levels to display of the palette.}

\item{builtin}{Should the default R palettes be shown?}
}
\description{
Creates a plot with various color palettes.
}
\examples{
pal.chips(list(cubehelix, gnuplot, jet, tol.rainbow, tol,inferno, magma,plasma, viridis, parula), n=31)

}

