% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.compress}
\alias{pal.compress}
\title{Compress a color palette function to fewer colors}
\usage{
pal.compress(pal, n = 5, thresh = 2.5)
}
\arguments{
\item{pal}{A palette function or a vector of colors.}

\item{n}{Initial number of colors to use.}

\item{thresh}{Maximum allowable Lab distance from original palette}
}
\value{
A vector of colors
}
\description{
Compress a color palette function to fewer colors
}
\details{
Many color palette functions are defined with more colors than needed.
This function compresses a color palette function down to a sample
of colors that can be passed into 'colorRampPalette' and re-create the
original palette with a just-noticeable-difference.

Color palettes that are defined as a smoothly varying ramp between a set of
colors often compress quite well.
Color palettes that are defined by functions may not compress well.
See the examples.
}
\examples{
# The 'cm.colors' palette compresses to only 3 colors
cm2 <- pal.compress(cm.colors, n=3)
pal.bands(cm.colors(255), colorRampPalette(cm2)(255), cm2)
title("cm.colors")

# The 'heat.colors' palette needs 84 colors
heat2 <- pal.compress(heat.colors, n=3)
pal.bands(heat.colors(255), colorRampPalette(heat2)(255), heat2)
title("heat.colors")

# The 'topo.colors' palette needs 249 colors because of the discontinuity
# topo2 <- pal.compress(topo.colors, n=3)
# pal.bands(list(topo.colors(255), colorRampPalette(topo2)(255), topo2))
# title("topo.colors")

# palettes that are defined by functions are sometimes difficult to compress
# into a colorRampPalette with a small number of colors
pal.bands(terrain.colors(11), terrain.colors(21), terrain.colors(31),
          terrain.colors(41), terrain.colors(51),
          cm.colors(11), cm.colors(21), cm.colors(31), cm.colors(41), cm.colors(51),
          labels=c("11","21","31","41","51","11","21","31","41","51"))

# smooth palettes are easier to compress
p1 <- coolwarm(255)
p2 <- colorRampPalette( coolwarm(15) )(255)
pal.bands(p1,p2,coolwarm(15), labels=c("coolwarm","colorRampPalette 15","15"))
pal.maxdist(p1,p2) # 2.33
 
}
\author{
Kevin Wright
}
\references{
None
}

