% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.heatmap}
\alias{pal.heatmap}
\title{Show a palette of colors with a heatmap}
\usage{
pal.heatmap(pal, n = 25, miss = 0.05)
}
\arguments{
\item{pal}{A palette function or a vector of colors.}

\item{n}{The number of colors to display for palette functions.}

\item{miss}{Fraction of squares with missing values, default .05.}
}
\value{
None.
}
\description{
A random heatmap is generated (with 5% missing values) and a key is added
to the heatmap by appending a blank column and then a column with the
palette colors.
}
\details{
What to look for:

1. Can the value of each cell be correctly interpreted using the key on 
the right side?

2. Can missing be identified?
}
\examples{
pal.heatmap(brewer.paired(12), n=12)
pal.heatmap(coolwarm(12), n=12)
pal.heatmap(tol, n=12)
pal.heatmap(glasbey)
pal.heatmap(kelly, n=22)

}
\author{
Kevin Wright
}
\references{
None
}

