% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.scatter}
\alias{pal.scatter}
\title{Test a palette of colors with a scatterplot}
\usage{
pal.scatter(pal, n)
}
\arguments{
\item{pal}{A palette function or a vector of colors.}

\item{n}{The number of colors to display for palette functions.}
}
\value{
None.
}
\description{
Details.
}
\examples{
pal.scatter(glasbey, n=31)
}
\author{
Kevin Wright
}
\references{
None.
}

