% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\name{pal.test}
\alias{pal.test}
\title{Create a test image of a color palette.}
\usage{
pal.test(pal)
}
\arguments{
\item{pal}{A palette function.}
}
\value{
None.
}
\description{
1. Z-order curve.  See: https://en.wikipedia.org/wiki/Z-order_curve
A good color palette should be able to resolve 4 sub-squares within
each of the 16 squares.
}
\details{
2. Contrast Sensitivity Function. Also called a contrast-resolution grating.
Frequency increases from left to right, contrast decreases from bottom
to top.  You can also think of this as a sine wave with periodicity becoming
shorter at the right side and the amplitude decreasing to 0 at the top.
Poor colormaps hide variation at the top.
See: http://blog.kasson.com/?m=201404.

3. Frequency ramp. See: http://inversed.ru/Blog_2.htm
Are the vertical bands visible across the full vertical axis?

4. 5. Two images of the 'volcano' elevation data in R using forward/reverse
colors. Try to find the highest point on the volcano peak.  Many palettes
with dark colors at one end of the palette hide the peak (e.g. viridis).
Also try to decide if the upperleft and upperright corners are the same color.
}
\examples{
pal.test(parula)
}
\references{
# See links inline.
}

