% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.maxcolors.R
\name{pals.maxcolors}
\alias{pals.maxcolors}
\title{Pals max colors}
\usage{
pals.maxcolors()
}
\value{
A data frame with the maximum number of colors for each palette.
}
\description{
This function returns a data frame with the maximum number of colors for
 each palette currently available within the \pkg{pals} package.
}
\examples{
dat <- pals.maxcolors()
}
\author{
R code by Brian M Schilder.
}
