% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\docType{package}
\name{pals}
\alias{pals}
\alias{pals-package}
\title{pals: A package for comprehensive palettes and palette evaluation tools}
\description{
pals: A package for comprehensive palettes and palette evaluation tools
}
\details{
The design goals of the package are:
\itemize{

\item
All palettes are functions that return a vector of 'n' colors.

\item
The palette function names use only lowercase letters.

\item
The 'data' directory is not used.

\item
No dependencies/imports of palettes.

\item
Extensive collection of palettes.

\item
Be memory efficient.

\item
Multiple tools to evaluate palettes.
}

To learn more, see the vignettes:
\code{browseVignettes(package="pals")}
}

