% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viridis.r
\docType{import}
\name{reexports}
\alias{inferno}
\alias{magma}
\alias{plasma}
\alias{reexports}
\alias{viridis}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{viridisLite}{\code{\link[viridisLite]{viridis}}, \code{\link[viridisLite]{magma}}, \code{\link[viridisLite]{inferno}}, \code{\link[viridisLite]{plasma}}}
}}
\keyword{internal}

