% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdzjoint.R
\name{stdzjoint}
\alias{stdzjoint}
\title{Standardize the input matrix}
\usage{
stdzjoint(X)
}
\arguments{
\item{X}{A matrix of any dimension}
}
\value{
Returns a joint standardized matrix
}
\description{
Standardize the input matrix
}
