% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{discrete}
\alias{alphabet}
\alias{cols25}
\alias{discrete}
\alias{glasbey}
\alias{kelly}
\alias{stepped}
\alias{tol}
\title{Discrete palettes}
\usage{
alphabet(n = 26)

cols25(n = 25)

glasbey(n = 32)

kelly(n = 22)

stepped(n = 20)

tol(n)
}
\arguments{
\item{n}{Number of colors to return.}
}
\value{
A vector of colors.
}
\description{
Color palettes designed for discrete, categorical data with a small number of
categories.
}
\details{
The \code{alphabet} palette has 26 distinguishable colors that have logical names
starting with the letters A, B, ... Z.
This palette is based on the work by Green-Armytage (2010), but uses the
names 'orange' instead of 'orpiment', and 'magenta' instead of 'mallow'.

The \code{cols25} palette was created experimentally by Wright (unpublished)
to create a set of colors that are distinct.

The \code{glasbey} palette by Glasbey et al (2007) has 32 colors that are
maximally distinct. Glasbey has 'white' as the second color, but in this
version of the palette, the color 'white' is moved to the end, and is
actually light-gray, #F2F3F4.
There exists a 256-color palette palette of 'distinct' colors.
We think this is absurd--even 32 colors are hard to distinguish.

The \code{kelly} palette of 22 colors maximize the contrast
between colors in a set if the colors are chosen in sequential order.
Kelly paid attention to the needs of people with color blindness.  The
first nine colors work well for such people and people with normal vision.
Kelly did not provide RGB color values, and the paper was in black-and-white.
A color image of the Kelly palette can be found in Green-Armytage (2010).
The color values used here are from
http://hackerspace.kinja.com/iscc-nbs-number-hex-r-g-b-263-f2f3f4-242-243-244-267-22-1665795040
The color 'white' has been re-defined as light-gray, #F2F3F4.
Commentary: We think the kelly palette has an over-abundance of orange-ish
colors, the purples are not very distinct, etc.  See the pal.heatmap
example.

The \code{stepped} palette has 20 colors (5 hues, 4 levels within each hue)
that is useful for showing varying levels within categories.
In order to better separate these colors in RGB space, we
moved red hue 0 to hue 350, green hue 80 to hue 90

The \code{tol} palette has 12 colors by Paul Tol.
Any printer that conforms to international standard ISO 12647-2 can
reproduce all colours.
}
\examples{

# ----- alphabet -----
alphabet()
alphabet()["jade"]
pal.cube(alphabet)
pal.bands(alphabet,n=26)
pal.map(alphabet)
pal.heatmap(alphabet)

# ----- cols25 -----
pal.bands(cols25,n=25)
pal.heatmap(cols25)

# ------ glasbey ------
pal.bands(glasbey,n=32)
pal.cube(glasbey, n=32) # Blues are close together
pal.heatmap(glasbey(32))

# ----- kelly -----
pal.bands(kelly,n=22)
pal.heatmap(kelly(22))

# ----- stepped -----
pal.bands(stepped,n=20)
pal.heatmap(stepped, n=20)

# ----- tol -----
pal.bands(tol,n=12)
pal.heatmap(tol, 12)

}
\author{
Palette colors by various authors. R code by Kevin Wright.
}
\references{
Chris Glasbey, Gerie van der Heijden, Vivian F. K. Toh, Alision Gray (2007).
Colour Displays for Categorical Images.
\emph{Color Research and Application}, 32, 304-309.
http://doi.org/10.1002/col.20327

P. Green-Armytage (2010): A Colour Alphabet and the Limits of Colour Coding.
\emph{Colour: Design & Creativity} (5) (2010): 10, 1-23.
www.aic-color.org/journal/v5/jaic_v5_06.pdf

K. Kelly (1965): Twenty-two colors of maximum contrast.
\emph{Color Eng}., 3(6), 1965.
http://www.iscc.org/pdf/PC54_1724_001.pdf

Paul Tol (2012). Color Schemes. SRON technical note, SRON/EPS/TN/09-002.
https://personal.sron.nl/~pault/

Color Schemes Appropriate for Scientific Data Graphics
http://geog.uoregon.edu/datagraphics/color_scales.htm
}

