% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.bands}
\alias{pal.bands}
\title{Show palettes as colored bands}
\usage{
pal.bands(..., n = 100, labels = NULL, title = NULL, gap = 0.1)
}
\arguments{
\item{...}{Palettes, either functions or vectors of colors.}

\item{n}{The number of colors to display for palette functions.}

\item{labels}{Labels for palettes}

\item{title}{Title at top of page.}

\item{gap}{Vertical gap between bars, default is 0.1}
}
\description{
Show palettes as colored bands.
}
\details{
What to look for:

1. A good discrete palette has distinct colors.

2. A good continuous colormap does not show boundaries between colors.
The rainbow() palette is poor, showing bright lines at yellow, cyan, pink.
}
\examples{
pal.bands(cubehelix, gnuplot, jet, tol.rainbow, inferno, magma, plasma, viridis, parula, gap=.05)
pal.bands(c('red','white','blue'), rainbow)

}

