% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.bands}
\alias{pal.bands}
\title{Show palettes as colored bands}
\usage{
pal.bands(..., n = 100, labels = NULL, title = NULL)
}
\arguments{
\item{...}{Palettes, either functions or vectors of colors.}

\item{n}{The number of colors to display for palette functions.}

\item{labels}{Labels for palettes}

\item{title}{Title at top of page.}
}
\description{
Show palettes as colored bands.
}
\examples{
pal.bands(cubehelix, gnuplot, jet, tol.rainbow, inferno, magma, plasma, viridis, parula, n=51)
pal.bands(c('red','white','blue'), rainbow)

}

