% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.cube}
\alias{pal.cube}
\title{Show a palette of colors in three dimensional RGB or LUV space}
\usage{
pal.cube(pal, n = 100, label = FALSE, type = "RGB")
}
\arguments{
\item{pal}{A palette function or a vector of colors.}

\item{n}{The number of colors to display for palette functions.}

\item{label}{If TRUE, show color name/value on plot}

\item{type}{Either "RGB" (default) or "LUV".}
}
\value{
None
}
\description{
The supplied color palette is converted to red,green,blue values
and plotted in a three-dimensional scatterplot.
}
\examples{
\dontrun{
pal.cube(cubehelix)
pal.cube(glasbey, n=32) # blues are too close to each other
pal.cube(glasbey, n=32, type="LUV")
pal.cube(cols25(25), type="LUV", label=TRUE)
# To open a second cube
rgl.open() # Open a new RGL device
rgl.bg(color = "white") # Setup the background color
pal.cube(colors()[c(1:152, 254:260, 362:657)]) # All R non-grey colors
}

}
\references{
None
}

