% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals.R
\name{pal.heatmap}
\alias{pal.heatmap}
\title{Test palette effectiveness with a heatmap}
\usage{
pal.heatmap(pal, n = 25)
}
\arguments{
\item{pal}{A palette function/vector.}

\item{n}{Number of colors in the palette.}
}
\value{
None.
}
\description{
A random heatmap is generated (with 5% missing values) and a key is added
to the heatmap by appending a blank column and then a column with the
palette colors.
}
\details{
In an effective palette, the value of each cell can be correctly
interpreted using the key on the right side, and missing values
are identified.
}
\examples{
pal.heatmap(paired(12), n=12)
pal.heatmap(coolwarm(12), n=12)
pal.heatmap(tol)
pal.heatmap(glasbey)
pal.heatmap(kelly)
}
\author{
Kevin Wright
}
\references{
None
}

