% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.heatmap}
\alias{pal.heatmap}
\title{Show a palette/colormap with a heatmap}
\usage{
pal.heatmap(pal, n = 25, miss = 0.05, main = "")
}
\arguments{
\item{pal}{A palette function or a vector of colors.}

\item{n}{The number of squares vertically in the heatmap.}

\item{miss}{Fraction of squares with missing values, default .05.}

\item{main}{Main title}
}
\value{
None.
}
\description{
Show a palette/colormap with a heatmap
}
\examples{
pal.heatmap(brewer.paired, n=12)
pal.heatmap(coolwarm, n=12)
pal.heatmap(tol, n=12)
pal.heatmap(glasbey, n=32)
pal.heatmap(kelly, n=22)

}
\author{
Kevin Wright
}
\references{
None
}

