% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{pal.safe}
\alias{pal.safe}
\title{Test safety of palette for black/white, colorblind}
\usage{
pal.safe(pal, n = 25)
}
\arguments{
\item{pal}{}

\item{n}{}
}
\value{
None.
}
\description{
.desc
}
\details{
.details
}
\examples{
pal.safe(glasbey(31))
pal.safe(rainbow(100))
pal.safe(cubicyf(21))
}
\author{
Kevin Wright
}
\references{
None
}

