% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\name{pal.scatter}
\alias{pal.scatter}
\title{Test palette effectiveness with a scatterplot}
\usage{
pal.scatter(pal, n)
}
\arguments{
\item{pal}{}

\item{n}{}
}
\value{
None.
}
\description{
Details.
}
\examples{
pal.scatter(glasbey, n=31)
}
\author{
Kevin Wright
}
\references{
None.
}

