% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{pal.test}
\alias{pal.test}
\title{Show a palette of colors with multiple images}
\usage{
pal.test(pal, title = substitute(pal))
}
\arguments{
\item{pal}{A palette function or a vector of colors.}

\item{title}{Title to display at the top of the test image}
}
\value{
None.
}
\description{
1. Z-curve
}
\details{
2. Contrast Sensitivity Function.

3. Frequency ramp. See: http://inversed.ru/Blog_2.htm
Are the vertical bands visible across the full vertical axis?

4. 5. Two images of the 'volcano' elevation data in R using forward/reverse
colors. Try to find the highest point on the volcano peak.  Many palettes
with dark colors at one end of the palette hide the peak (e.g. viridis).
Also try to decide if the upperleft and upperright corners are the same color.

6. Luminosity in red, green, blue, and grey.
}
\examples{
pal.test(parula)
pal.test(viridis)
pal.test(coolwarm)

}
\author{
Kevin Wright
}
\references{
# See links above.
}

