% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.r
\docType{package}
\name{pals-package}
\alias{pals-package}
\title{Comprehensive color palettes and evaluation tools}
\description{
Design goals:
}
\details{
1. All palettes are functions that return a vector of 'n' colors.

2. The function names use only lowercase letters.

3. The 'data' directory is not used.
}
\author{
Kevin Wright
}

